arm: dts: artik310s: modify shift of SD
This patch modifies drive_shift and sample_shift of SD.

The drive_shift can have values as 0 to 3. The value "0" means
degree 0, "1" means degree 90, "2" means degree 180 and "3" means
degree 270. In the same as drive_shift, the sample_shift can have
values as 0 to 3. The meanings of values are also the same.

There is sometimes SD init fail during booting, and it occured
because of command timeout. This patch solves it.

Change-Id: I0cd15bc6e92317d027066d74a4dface807eefd35
Signed-off-by: Sowon Na <sowon.na@samsung.com>