artik310s: configs: enable eMMC hwreset
This patch is able to enable hwreset of eMMC during eMMC format.
"mmc hwreset enable /dev/mmcblk0" command will will RST_n signal can be
permanently enabled and the control is located in EXTCSD[162]. This
will be never changed once it's written. Belows are the meaning of the
RST_n_Function[162] and we'll always make it can be enabled.

Bit[1:0]: RST_n_ENABLE (Readable and Writable once)
0x0: RST_n signal is temporarily disabled (default)
0x1: RST_n signal is permanently enabled
0x2: RST_n signal is permanently disabled
0x3: Reserved

Change-Id: I3ac764a92c605bb5086b9ebe99b32a991ab6f2ae
Signed-off-by: Chanho Park <chanho61.park@samsung.com>