net: add nexell dwc eqos driver
This patch adds Nexell Synopsys Designware QOS driver which is based on
dwc_eth_qos driver. The nxp3220 SoC has two ethernet MACs(mac0 is GMAC
and mac1 is EMAC). Unlike tegra186, the nxp3220 does not need a RX clock
of GMAC. It is sufficient to provide TX clock for GMAC.

Change-Id: Id76460a184f25283b29ee59aeae73bc6e487770b
Signed-off-by: Chanho Park <chanho61.park@samsung.com>