board: nexell: artik310s_trike: add board type check
artik310s module supports 3-types of sub-board. They're Dev-Kit,
Eagleye310 and Full-Function board. To check the board type at boot
time, read the ADC-2 value to identify the board.

0~0.1v = Eagleye310
0.4~0.5v = Full-Function
1.5 ~ 1.8v = Dev-Kit

To be more configurable this check, I added a simple "artik,boardtype"
compatible and it will get a en-gpio which enables the ADC2 and an adc
node with channel number.

Change-Id: I1d06aa36bd9b654b6b83f707a8557ba13998be19
Signed-off-by: Jaewon Kim <jaewon02.kim@samsung.com>
Signed-off-by: Chanho Park <chanho61.park@samsung.com>