board: nexell: add boot_mode driver
This patch adds nexell,boot_mode driver in order to hook reboot with an
arg such as bootloader, recovery and so on.

The "syscon-reboot-mode" driver of linux kernel will set an alive register
which is able to keep its value during soft reset. This driver will
check a value of register and enter proper modes.

Change-Id: I3b8dcfd2541904b1214efbfc96370b464de4d2be
Signed-off-by: Chanho Park <chanho61.park@samsung.com>